backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-filter==24.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
psycopg[binary]==3.2.1
